Create function for close and open
