+++ src/functions/graphql/entities/election/index.js
