+++ src/functions/graphql/entities/election/typeDef.graphql
