+++ src/functions/graphql/entities/vote/index.js
