+++ src/functions/graphql/entities/vote/typeDef.graphql
