+++ src/functions/graphql/protectedResolver.js
